strip the original script + little fix

+++ obfuscate.py
@@ -5,23 +5,36 @@ from Crypto.Random import get_random_bytes
 from strip import strip
 import os
 
+def get_code(filename):
+    # return the code from a python script
+    with open(filename, 'r') as f:
+        code = f.read()
+    return code
 
-def get_bytecode(filename):
-    # Return bytecode from a pythonn script
+def get_bytecode(code):
+    # Return bytecode from a python script
     i = 0
-    tmpfilename = '.tmp'
-    while os.path.exists(tmpfilename+'.pyc'):
-        tmpfilename = '.tmp' + str(i)
-    tmpfilename = tmpfilename + '.pyc'
+    tmp = '.tmp.py'
+    while os.path.exists(tmp):
+        tmp = f'.tmp{i}.py'
+        i += 1
 
-    with open (py_compile.compile('test.py', cfile=tmpfilename), "rb") as f:
-        bytecode = f.read()
-    os.remove(tmpfilename)
-    return bytecode
+    tmpfilename = tmp + 'c'
+    try:
+        with open(tmp, 'w') as f:
+            f.write(code)
+        with open(py_compile.compile(tmp, cfile=tmpfilename), "rb") as f:
+            bytecode = f.read()
+        return bytecode
+    finally:
+        if os.path.exists(tmp):
+            os.remove(tmp)
+        if os.path.exists(tmpfilename):
+            os.remove(tmpfilename)
 
 def crypto_checksum(bytecode):
-    # Return a python script wich perform a Blake3 checksum
-    # of the bytecode and encrypt it with ChaCha20-Poly1305
+    # Return a python script which performs a Blake3 checksum
+    # of the bytecode and encrypts it with ChaCha20-Poly1305
     checksum = blake3(bytecode).hexdigest()
     key = get_random_bytes(32)
     nonce = get_random_bytes(24)
@@ -38,8 +51,9 @@ if blake3(plaintext).hexdigest() == '{checksum}':
     return code
 
 def obfuscate(filename):
-    # TODO: strip before getting the bytecode, may do a tempfile (be sure the file does not exist)
-    bytecode = get_bytecode(filename)
+    code = get_code(filename)
+    code = strip(code)
+    bytecode = get_bytecode(code)
     code = crypto_checksum(bytecode)
     code = strip(code)
     return code
